requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2
djangorestframework==3.16.0
idna==3.10
mysqlclient==2.2.7
pillow==11.2.1
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
